+++ ts/features/newWallet/navigation/params.ts
@@ -1,0 +1,5 @@
+import { WalletRoutes } from "./routes";
+
+export type WalletParamsList = {
+  [WalletRoutes.WALLET_CARD_ONBOARDING]: undefined;
+};
