+++ ts/navigation/params/AppParamsList.ts
@@ -44,10 +44,8 @@ import {
 } from "../../features/idpay/unsubscription/navigation/navigator";
 import { MessagesParamsList } from "../../features/messages/navigation/params";
 import { MESSAGES_ROUTES } from "../../features/messages/navigation/routes";
-import {
-  WalletParamsList as NewWalletParamsList,
-  WalletRoutes as NewWalletRoutes
-} from "../../features/newWallet/navigation";
+import { WalletParamsList as NewWalletParamsList } from "../../features/newWallet/navigation/params";
+import { WalletRoutes as NewWalletRoutes } from "../../features/newWallet/navigation/routes";
 import { PaymentsBarcodeParamsList } from "../../features/payments/barcode/navigation/params";
 import { PaymentsBarcodeRoutes } from "../../features/payments/barcode/navigation/routes";
 import { PaymentsCheckoutParamsList } from "../../features/payments/checkout/navigation/params";
